Create DevilRat_eyesmouth
This is for a Halloween Decoration in which I take an Adafruit Gemma, wire it to 3 Adafruit Floras and a 5-volt DC power supply and place the assembly a LARGE plastic rat. Two of the Floras are the eyes, and one is the mouth. I attached the Large rat to a a belt-driven gantry and had it drive the length of the gantry, approaching our visitor with bright, scary eyes. The code comes from Adafruit.com. I've made minimal changes -- MiketheChap